Cleanup build failures using c6x gcc
Finally, the kernel is buildable by either the GNU tools or the
Texas Instruments tools. Toolchain selection is done through the
kernel configuration (CONFIG_TI_C6X_LINKER,CONFIG_TI_C6X_COMPILER).

+++ arch/c6x/include/asm/ptrace.h
@@ -196,9 +196,9 @@
 #ifndef __ASSEMBLY__
 
 #ifdef _BIG_ENDIAN
-#define REG_PAIR(odd,even) int odd; int even
+#define REG_PAIR(odd,even) unsigned long odd; unsigned long even
 #else
-#define REG_PAIR(odd,even) int even; int odd
+#define REG_PAIR(odd,even) unsigned long even; unsigned long odd
 #endif
 
 /*
